+++ packages/migration/CHANGELOG.md
@@ -4,6 +4,9 @@
 
 - added `--dry-run` option to run pending migrations without commiting them to quickly check for errors
 
+- Dependencies bump:
+  - `@fewlines/fwl-database`: `1.1.0` => `@fwl/database`: `0.1.0` (Package was renamed)
+
 ## 0.1.5 - 2020-07-27
 
 - Dependencies bump: